@@ -1,12 +1,12 @@
 name: absinthe_socket
 description: A GraphQL subscription client for Absinthe sockets (cloned for Postagraph).
-version: 0.0.3
+version: 0.0.4
 author: Josh Adams <josh@smoothterminal.com>
 homepage: http://github.com/dailydrip/absinthe_socket
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   # Talk to phoenix channels
-  phoenix_wings: ^0.2.1
+  phoenix_wings: ^1.0.2
